feat: Search functionality added
